refactor: replace classes
