+++ src/main/java/ru/yandex/practicum/filmorate/controller/UserController.java
@@ -4,7 +4,6 @@ import lombok.AllArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.service.UserService;
-import ru.yandex.practicum.filmorate.service.Validator;
 
 import javax.validation.Valid;
 import java.util.Collection;
