+++ src/main/java/ru/yandex/practicum/filmorate/storage/FilmStorage.java
@@ -5,7 +5,7 @@ import ru.yandex.practicum.filmorate.model.Film;
 import java.util.Collection;
 
 public interface FilmStorage {
-    Film createFilm(Film film);
+    Film saveFilm(Film film);
 
     Film updateFilm(Film film);
