+++ src/main/java/ru/yandex/practicum/filmorate/storage/UserStorage.java
@@ -5,7 +5,7 @@ import ru.yandex.practicum.filmorate.model.User;
 import java.util.Collection;
 
 public interface UserStorage {
-    User createUser(User user);
+    User saveUser(User user);
 
     User updateUser(User user);
