+++ src/main/java/ru/yandex/practicum/filmorate/storage/impl/InMemoryFilmStorage.java
@@ -15,7 +15,7 @@ public class InMemoryFilmStorage implements FilmStorage {
     private final Map<Integer, Film> films = new HashMap<>();
 
     @Override
-    public Film createFilm(Film film) {
+    public Film saveFilm(Film film) {
         films.put(film.getId(), film);
         return film;
     }
